Adicionado na classe Book o atributo ID, não havia sido inserido e devido a isso não tinha como informar o ID no método UpdateBook. Isso na classe STARTED ocasionava uma falha que sempre atualizava o primeiro registro da coluna do banco, não havia como escolher qual linha seria alterada.
Foi criado dentro do pacote Domain o pacote "Book" para que fosse separado as classes referente a Book e seu gerenciamento. No caso classes BOOK e BOOKMANAGER.

(The ID attribute was added to the Book class, it had not been added before, and because of this, it was not possible to inform the ID in the UpdateBook method. This caused a failure in the STARTED class that always updated the first record of the database column, and there was no way to choose which row to change.A 'Book' package was created within the Domain package to separate the classes related to Book and its management. In this case, the BOOK and BOOKMANAGER classes.)